refactor: 패키지 구조 변경
클래스들을 용도 별로 구분해 패키지 구조를 구성함.
