+++ src/main/java/lotto/constant/Constants.java
@@ -1,4 +1,4 @@
-package lotto;
+package lotto.constant;
 
 public final class Constants {
