+++ src/main/java/lotto/domain/LottoRank.java
@@ -1,8 +1,8 @@
-package lotto.service;
+package lotto.domain;
 
-import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.List;
+import lotto.constant.Error;
 
 public enum LottoRank {
