+++ src/main/java/lotto/service/LottoComparator.java
@@ -2,7 +2,9 @@ package lotto.service;
 
 import java.util.ArrayList;
 import java.util.List;
-import lotto.Lotto;
+import lotto.domain.Lotto;
+import lotto.domain.LottoRank;
+import lotto.domain.Standard;
 
 public class LottoComparator {
