+++ src/test/java/lotto/LottoComparatorTest.java
@@ -4,9 +4,12 @@ package lotto;
 import static org.assertj.core.api.Assertions.*;
 
 import java.util.List;
+import lotto.domain.Lotto;
+import lotto.domain.LottoNumber;
+import lotto.domain.WinningNumbers;
 import lotto.service.LottoComparator;
-import lotto.service.LottoRank;
-import lotto.service.Standard;
+import lotto.domain.LottoRank;
+import lotto.domain.Standard;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
