+++ src/test/java/lotto/LottoMakerTest.java
@@ -3,6 +3,7 @@ package lotto;
 import static org.assertj.core.api.Assertions.*;
 
 import java.util.List;
+import lotto.domain.Lotto;
 import lotto.service.LottoMaker;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
