+++ src/test/java/lotto/LottoPurchaserTest.java
@@ -3,6 +3,8 @@ package lotto;
 import static org.assertj.core.api.Assertions.*;
 
 import java.util.List;
+import lotto.domain.Lotto;
+import lotto.domain.Price;
 import lotto.service.LottoPurchaser;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
