+++ src/test/java/lotto/LottoTest.java
@@ -1,6 +1,6 @@
 package lotto;
 
-import org.junit.jupiter.api.Disabled;
+import lotto.domain.Lotto;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
