+++ src/test/java/lotto/MessageTest.java
@@ -3,9 +3,10 @@ package lotto;
 import static org.assertj.core.api.Assertions.assertThat;
 
 import java.util.List;
+import lotto.domain.Lotto;
 import lotto.messge.IssueMessage;
 import lotto.messge.ReportMessage;
-import lotto.service.LottoRank;
+import lotto.domain.LottoRank;
 import lotto.service.WinningHistory;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
