+++ src/test/java/lotto/WinningHistoryTest.java
@@ -3,7 +3,7 @@ package lotto;
 import static org.assertj.core.api.Assertions.*;
 
 import java.util.List;
-import lotto.service.LottoRank;
+import lotto.domain.LottoRank;
 import lotto.service.WinningHistory;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
