+++ src/test/java/lotto/WinningNumbersTest.java
@@ -3,6 +3,8 @@ package lotto;
 import static org.assertj.core.api.Assertions.*;
 
 import java.util.List;
+import lotto.domain.LottoNumber;
+import lotto.domain.WinningNumbers;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
